functions/dha-queryPatientMedications/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.27.0
flask==2.*